requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
supabase==2.10.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pydantic-settings==2.7.0
httpx==0.28.1
python-dotenv==1.0.1
openpyxl==3.1.5
fpdf2==2.8.1
